requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.4
requests==2.32.3
Flask-WTF==1.2.1
PyMySQL==1.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.0